pyproject.toml
[project]
name = "openhands-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openhands-sdk>=1.11.4",
    "openhands-tools>=1.11.4",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
